Rename the directory for Sweeping.
